pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "face-detect"
version = "0.2.0"
description = "Arquive — personal media archive with face detection, adaptive streaming & live TV"
requires-python = ">=3.10,<3.14"
dependencies = [
    # --- Existing (face detection) ---
    "insightface>=0.7",
    "onnxruntime-gpu>=1.17",
    "opencv-python>=4.8",
    "faiss-cpu>=1.7",
    "flask>=3.0",
    "pyyaml>=6.0",
    "numpy>=1.24",
    "jinja2>=3.1",
    "requests>=2.31",
    # --- New (Arquive) ---
    "bcrypt>=4.0",
    "PyJWT>=2.8",
    "cryptography>=41.0",
    "Pillow>=10.0",
    "PyPDF2>=3.0",
    "ipytv>=0.2",
]

[project.scripts]
face-detect = "face_detect.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
